Create CODE-1
This program uses the Wire and LiquidCrystal_I2C libraries to communicate with the LED display over the I2C bus. The display is initialized and the backlight is turned on in the setup function. In the loop function, the display is cleared and two special messages for New Year's are displayed on the first and second lines of the display. The program then waits for 5 seconds before displaying the next message.

You can customize the messages and the delay time to suit your needs. You can also use the lcd.scrollDisplayLeft() and lcd.scrollDisplayRight() functions to scroll the messages across the display, or use other functions provided by the LiquidCrystal_I2C library to customize the appearance of the messages.